// src/generated/t_tech/invest/grpc/instruments.ts
export interface GetFuturesMarginResponse {
  /** Гарантийное обеспечение при покупке. */
  initialMarginOnBuy:
    | MoneyValue
    | undefined;
  /** Гарантийное обеспечение при продаже. */
  initialMarginOnSell:
    | MoneyValue
    | undefined;
  /** Шаг цены. */
  minPriceIncrement:
    | Quotation
    | undefined;
  /** Стоимость шага цены. */
  minPriceIncrementAmount: Quotation | undefined;
}
// ...
export const GetFuturesMarginResponse: MessageFns<GetFuturesMarginResponse> = {
  encode(message: GetFuturesMarginResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.initialMarginOnBuy !== undefined) {
      MoneyValue.encode(message.initialMarginOnBuy, writer.uint32(10).fork()).join();
    }
    if (message.initialMarginOnSell !== undefined) {
      MoneyValue.encode(message.initialMarginOnSell, writer.uint32(18).fork()).join();
    }
    if (message.minPriceIncrement !== undefined) {
      Quotation.encode(message.minPriceIncrement, writer.uint32(26).fork()).join();
    }
    if (message.minPriceIncrementAmount !== undefined) {
      Quotation.encode(message.minPriceIncrementAmount, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetFuturesMarginResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetFuturesMarginResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.initialMarginOnBuy = MoneyValue.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.initialMarginOnSell = MoneyValue.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.minPriceIncrement = Quotation.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.minPriceIncrementAmount = Quotation.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetFuturesMarginResponse {
    return {
      initialMarginOnBuy: isSet(object.initialMarginOnBuy)
        ? MoneyValue.fromJSON(object.initialMarginOnBuy)
        : isSet(object.initial_margin_on_buy)
        ? MoneyValue.fromJSON(object.initial_margin_on_buy)
        : undefined,
      initialMarginOnSell: isSet(object.initialMarginOnSell)
        ? MoneyValue.fromJSON(object.initialMarginOnSell)
        : isSet(object.initial_margin_on_sell)
        ? MoneyValue.fromJSON(object.initial_margin_on_sell)
        : undefined,
      minPriceIncrement: isSet(object.minPriceIncrement)
        ? Quotation.fromJSON(object.minPriceIncrement)
        : isSet(object.min_price_increment)
        ? Quotation.fromJSON(object.min_price_increment)
        : undefined,
      minPriceIncrementAmount: isSet(object.minPriceIncrementAmount)
        ? Quotation.fromJSON(object.minPriceIncrementAmount)
        : isSet(object.min_price_increment_amount)
        ? Quotation.fromJSON(object.min_price_increment_amount)
        : undefined,
    };
  },

  toJSON(message: GetFuturesMarginResponse): unknown {
    const obj: any = {};
    if (message.initialMarginOnBuy !== undefined) {
      obj.initialMarginOnBuy = MoneyValue.toJSON(message.initialMarginOnBuy);
    }
    if (message.initialMarginOnSell !== undefined) {
      obj.initialMarginOnSell = MoneyValue.toJSON(message.initialMarginOnSell);
    }
    if (message.minPriceIncrement !== undefined) {
      obj.minPriceIncrement = Quotation.toJSON(message.minPriceIncrement);
    }
    if (message.minPriceIncrementAmount !== undefined) {
      obj.minPriceIncrementAmount = Quotation.toJSON(message.minPriceIncrementAmount);
    }
    return obj;
  },

  create(base?: DeepPartial<GetFuturesMarginResponse>): GetFuturesMarginResponse {
    return GetFuturesMarginResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<GetFuturesMarginResponse>): GetFuturesMarginResponse {
    const message = createBaseGetFuturesMarginResponse();
    message.initialMarginOnBuy = (object.initialMarginOnBuy !== undefined && object.initialMarginOnBuy !== null)
      ? MoneyValue.fromPartial(object.initialMarginOnBuy)
      : undefined;
    message.initialMarginOnSell = (object.initialMarginOnSell !== undefined && object.initialMarginOnSell !== null)
      ? MoneyValue.fromPartial(object.initialMarginOnSell)
      : undefined;
    message.minPriceIncrement = (object.minPriceIncrement !== undefined && object.minPriceIncrement !== null)
      ? Quotation.fromPartial(object.minPriceIncrement)
      : undefined;
    message.minPriceIncrementAmount =
      (object.minPriceIncrementAmount !== undefined && object.minPriceIncrementAmount !== null)
        ? Quotation.fromPartial(object.minPriceIncrementAmount)
        : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}